+++ pkgs/test/test/runner/compiler_runtime_matrix_test.dart
@@ -149,6 +149,23 @@ void main() {
                   : null,
             );
           }
+
+          test(
+            'enables asserts',
+            () async {
+              await d.file('test.dart', _assertionsEnabledTest).create();
+              var test = await runTest(testArgs);
+
+              expect(
+                test.stdout,
+                emitsThrough(contains('+1: All tests passed!')),
+              );
+              await test.shouldExit(0);
+            },
+            skip: compiler == Compiler.cli
+                ? 'https://github.com/dart-lang/test/issues/2718'
+                : false,
+          );
         },
       );
     }
@@ -199,3 +216,15 @@ void main() async {
   stderr.writeln('world');
   test('success', () {});
 }''';
+
+final _assertionsEnabledTest = '''
+import 'package:test/test.dart';
+
+void main() {
+  test('asserts are enabled', () {
+    expect(() {
+      assert(false);
+    }, throwsA(isA<AssertionError>()));
+  });
+}
+''';